Set all the db mocks timestamps precision to millis If we have different length in the redis sorted set, the timeline is not accurate
