+++ docker/db-graph/mocks/files.cypher
@@ -1,6 +1,6 @@
-MERGE (f:File {id: "2ZK2H8P2T5NG0"}) SET f.id = "2ZK2H8P2T5NG0", f.uri = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/files/2ZK2H8P2T5NG0", f.owner_id = "y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy", f.indexed_at = 1724134095 , f.created_at = 1724134095, f.src = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/blobs/2ZK2H8NS5T4G0", f.size = 85154, f.name = "myfile1", f.content_type = "image/jpeg", f.urls = "{\"main\":\"y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/2ZK2H8P2T5NG0\"}";
-MERGE (f:File {id: "2ZKH7K7B2RY00"}) SET f.id = "2ZKH7K7B2RY00", f.uri = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/files/2ZKH7K7B2RY00", f.owner_id = "y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy", f.indexed_at = 1724134095 , f.created_at = 1724134095, f.src = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/blobs/2ZKH7K73XMWG0", f.size = 85154, f.name = "myfile2", f.content_type = "image/jpeg", f.urls = "{\"main\":\"y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/2ZKH7K7B2RY00\"}";
-MERGE (f:File {id: "2ZKH7K7M9G3G0"}) SET f.id = "2ZKH7K7M9G3G0", f.uri = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/files/2ZKH7K7M9G3G0", f.owner_id = "y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy", f.indexed_at = 1724134095 , f.created_at = 1724134095, f.src = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/blobs/2ZKH7K7M0RY00", f.size = 85154, f.name = "myfile3", f.content_type = "image/jpeg", f.urls = "{\"main\":\"y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/2ZKH7K7M9G3G0\"}";
-MERGE (f:File {id: "2ZK2RH7KN91G0"}) SET f.id = "2ZK2RH7KN91G0", f.uri = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/files/2ZK2RH7KN91G0", f.owner_id = "sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy", f.indexed_at = 1724134095 , f.created_at = 1724134095, f.src = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/blobs/2ZK2RH7D1E400", f.size = 85154, f.name = "myfile1", f.content_type = "image/jpeg", f.urls = "{\"main\":\"sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/2ZK2RH7KN91G0\"}";
-MERGE (f:File {id: "2ZK1VCJN4YE00"}) SET f.id = "2ZK1VCJN4YE00", f.uri = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/files/2ZK1VCJN4YE00", f.owner_id = "sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy", f.indexed_at = 1724134095 , f.created_at = 1724134095, f.src = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/blobs/2ZK1VCJBYN4G0", f.size = 85154, f.name = "myfile2", f.content_type = "image/jpeg", f.urls = "{\"main\":\"sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/2ZK1VCJN4YE00\"}";
-MERGE (f:File {id: "003286NSMY490"}) SET f.id = "003286NSMY490", f.uri = "pubky://4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro/pub/pubky.app/files/003286NSMY490", f.owner_id = "4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro", f.indexed_at = 1724134095 , f.created_at = 1724134095, f.src = "pubky://4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro/pub/pubky.app/blobs/0032803PZX8Y0", f.size = 85154, f.name = "myfile1", f.content_type = "image/jpeg", f.urls = "{\"main\":\"4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro/003286NSMY490\"}";
+MERGE (f:File {id: "2ZK2H8P2T5NG0"}) SET f.id = "2ZK2H8P2T5NG0", f.uri = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/files/2ZK2H8P2T5NG0", f.owner_id = "y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy", f.indexed_at = 1724134095000 , f.created_at = 1734134095000, f.src = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/blobs/2ZK2H8NS5T4G0", f.size = 85154, f.name = "myfile1", f.content_type = "image/jpeg", f.urls = "{\"main\":\"y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/2ZK2H8P2T5NG0\"}";
+MERGE (f:File {id: "2ZKH7K7B2RY00"}) SET f.id = "2ZKH7K7B2RY00", f.uri = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/files/2ZKH7K7B2RY00", f.owner_id = "y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy", f.indexed_at = 1724134094000 , f.created_at = 1744134095000, f.src = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/blobs/2ZKH7K73XMWG0", f.size = 85154, f.name = "myfile2", f.content_type = "image/jpeg", f.urls = "{\"main\":\"y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/2ZKH7K7B2RY00\"}";
+MERGE (f:File {id: "2ZKH7K7M9G3G0"}) SET f.id = "2ZKH7K7M9G3G0", f.uri = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/files/2ZKH7K7M9G3G0", f.owner_id = "y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy", f.indexed_at = 1724134093000 , f.created_at = 1754134095000, f.src = "pubky://y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/pub/pubky.app/blobs/2ZKH7K7M0RY00", f.size = 85154, f.name = "myfile3", f.content_type = "image/jpeg", f.urls = "{\"main\":\"y4euc58gnmxun9wo87gwmanu6kztt9pgw1zz1yp1azp7trrsjamy/2ZKH7K7M9G3G0\"}";
+MERGE (f:File {id: "2ZK2RH7KN91G0"}) SET f.id = "2ZK2RH7KN91G0", f.uri = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/files/2ZK2RH7KN91G0", f.owner_id = "sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy", f.indexed_at = 1724134092000 , f.created_at = 1764134095000, f.src = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/blobs/2ZK2RH7D1E400", f.size = 85154, f.name = "myfile1", f.content_type = "image/jpeg", f.urls = "{\"main\":\"sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/2ZK2RH7KN91G0\"}";
+MERGE (f:File {id: "2ZK1VCJN4YE00"}) SET f.id = "2ZK1VCJN4YE00", f.uri = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/files/2ZK1VCJN4YE00", f.owner_id = "sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy", f.indexed_at = 1724134091000 , f.created_at = 1774134095000, f.src = "pubky://sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/pub/pubky.app/blobs/2ZK1VCJBYN4G0", f.size = 85154, f.name = "myfile2", f.content_type = "image/jpeg", f.urls = "{\"main\":\"sfgetccnq7s3h57a7imf6n7k5fqxus33yg85f1ndhnrnofjdmhjy/2ZK1VCJN4YE00\"}";
+MERGE (f:File {id: "003286NSMY490"}) SET f.id = "003286NSMY490", f.uri = "pubky://4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro/pub/pubky.app/files/003286NSMY490", f.owner_id = "4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro", f.indexed_at = 1724134090000 , f.created_at = 1784134095000, f.src = "pubky://4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro/pub/pubky.app/blobs/0032803PZX8Y0", f.size = 85154, f.name = "myfile1", f.content_type = "image/jpeg", f.urls = "{\"main\":\"4snwyct86m383rsduhw5xgcxpw7c63j3pq8x4ycqikxgik8y64ro/003286NSMY490\"}";
