+++ tests/watcher/bookmarks/viewer.rs
@@ -68,7 +68,7 @@ async fn test_homeserver_viewer_bookmark() -> Result<()> {
     assert_eq!(viewer_bookmark.id, bookmark_id);
 
     // INDEX_OP: Assert if the event writes the indexes
-    let result_bookmarks = PostStream::get_bookmarked_posts(&viewer_id, None, None)
+    let result_bookmarks = PostStream::get_bookmarked_posts(&viewer_id, None, None, None, None)
         .await
         .unwrap();
